Also install Ollama server if not available.
Upon installation we check if Ollama is already installed and offer to install it otherwise since it is a critical dependency to run a model locally for cmd to interact with.

I currently verified functionality for macOS & will eventually test it out in the rest of the environments. The next step is to not only install it but pull a Deepseek model and kick off the server on user's behalf.